Fixes a testcase related to trusted boot in UEFI boot mode
This testcase is to test validate request when trusted boot
is requested along with UEFI boot mode.

In the original testcase, the code flow was failing in
iscsi_deploy.py:validate_image_properties() but was raising the
exception InvalidParameterValue as expected by testcase.
This false positive was leading to incorrect scenario testing.

In the environment where it failed, tox timed out while testcase
was trying to connect to glance server leading to testcase
failure.

In either case the test scenario is exercising wrong part of the
code and needs to be fixed.

Change-Id: If6ac867566a255a3032088ce7b487947536ef08b
Closes-Bug: #1473004