Allow inspector driver to work in standalone mode
Before the Ironic inspector driver always tried to get a token from Keystone,
this was making it impossible to use in standalone mode.  This patch is
fixing this assumption by checking the "auth_strategy" configuration option
and if the value is not "keystone" we won't try to get a token from Keystone.
In test add test_is_standalone and test_not_standalone, check the driver
won't try to get a token from Keystone when use in standalone.

Change-Id: I5cb31d2db787501792c02b797deb8a710aa7131f
Closes-Bug:#1461525