Allow upgrading shared lock to an exclusive one
New task method upgrade_lock is a simpler and nicer way of doing:

 if task.shared:
     task = task_manager.acquire(task.context, task.node.id)

This is useful for e.g. periodic tasks to start with a shared lock
and upgrade it if any changes are needed.

One more use case is to allow vendor passthru implementations
to decide if they actually need an exclusive lock instead of
just providing them it by default.

Both use cases are not covered by this change.

Change-Id: I2c019f2855b61330b5b18cee34daa5f399a41f2c