Clean up tftp files if agent deployed disk image
After a deployment is complete, the tftp boot files that allow
a node to boot into the agent, should be removed from the tftp
server specificially if a whole disk image is deployed as the
whole disk image should have sufficent infrastucture to boot
its self.

If this is not done and the files are left on disk, then a
server may end up booting the agent from the network instead
of the instance operating system from local storage if the
firmware does not honor or support setting the boot device.

Change-Id: Ib6caa445cd577e295c89e908650b29f5823523ed
Closes-Bug: #1476489