Don't prevent updates if power transition is in progress
Current behaviour results in 409 errors for innocent operations
while e.g. slow (I've seen up to 1.5 minutes) power operation is on.
This, for example, happens when Inspector is in use: the last action
it does is to power off the node. If it takes time, many operations
won't be possible on a node right after power action was requested.

The problem with this particular operation is that it fails with 409
without any retries on a server side.

This patch allows update request to reach conductor and to be retried
inside the task_manager.

Closes-Bug: #1478867
Change-Id: I43ea1b05b5021a658f47a6006ee9bb86018b3833