Clean up instance_uuid as part of the node's tear down
Prior to this patch Ironic would expect Nova to clean up the instance
association, but this assumption does not play well with the standalone
version of Ironic or cleaning.

This patch keeps a backward compatibility layer with old versions of Nova
that will try to clean up the instance_uuid field even if it's already
cleaned.

Depends-On: I0b1b710056d48c8b7bb2b46fdaba192922926420
Closes-Bug: #1436568
Change-Id: I58bc848ab3a60d3e9ab16d00ce96a603df1c244a