Add Nova scheduler_tracks_instance_changes config to docs
This option defaults to True and doesn't really make sense in an Ironic
context. It tracks instances by host in the scheduler, to allow for
better choosing of a host. However, Ironic only allows for one host, so
there's no opportunity to schedule to a host with more capacity.

Additionally, with the potential to have hundreds or thousands of
instances on a single "host" in Ironic, the get_by_host query this
config causes to run can completely knock over the scheduler.

Let's recommend turning it off. :)

Change-Id: I3c34dd8ef06a339606e56cd57d5b7d523dedc09b