Fix warnings on doc builds
In the sphinx config, mock out the nova imports we do for the
compute manager so the docs will build with fewer errors.

Exclude the alembic migration environment from autodoc.

Fix the remaining markup errors, which are now easier to spot
given the other fixes.

Note that this leaves one remaining warning, which is that the
generated API documentation isn't currently included in a
toctree. This is easy to fix but it would be good to discuss
whether/where to include it in the documentation hierarchy.

Closes-Bug: 1277282
Change-Id: Iee5fc3eeb95a09198d5234d0ea05e150ab26f16b