Use the oslo_utils.timeutils 'StopWatch' class
Instead of using time.time() to keep track of durations
the stop watch class provides this same functionality in a
nicely made helper class. It also internally uses monotonically
increasing time so that drifts/going backwards should not be
possible so it seems like a nice change to just use it vs computing
similar information.

Change-Id: Ic8906f42f6997412fffe950d4833dc2ad88ecf28