Remove CONF.agent.agent_pxe_bootfile_name
This commit removes CONF.agent.agent_pxe_bootfile_name as
it is not used anywhere. The usage of this CONF option was
removed in Id68d3a89cb72e11e8a52ea69e84816aea2cbf2b1. Even
before that, it wasn't used in the proper way during
deployment of nodes.  It was only used in take_over() for
AgentDeploy, for instances that always booted from disk.

Change-Id: Ib25ec6ec7cd900ee637bba035e0aeab90e9309c5