Update docstring for agent deploy's take_over
For IPA deploys, if a conductor takes over a node that has been
deployed by a different conductor, there is no need for that
conductor to do anything because only local boots are done.
This updates the docstring for the agent deploy's take_over()
to reflect this. For more details, see
Id68d3a89cb72e11e8a52ea69e84816aea2cbf2b1.

Change-Id: Ifb495d74c116abde853a1f029fae34e56ca264fc