Refactor the essential prop list of inspect driver
This patch is a groundwork for iRMC inspect driver [1], and refactors
the essential property list defined in iLO inspect driver into the common
base class.
This patch is also a reflection of the gerrit review comment [2] which
advised that the essential property list defined in iLO should be
removed in separate patch if you add the essential property list to
base.

[1] https://review.openstack.org/#/c/196480/
[2] https://review.openstack.org/#/c/196480/7/ironic/drivers/base.py

Change-Id: Iacce751bceb6ff8183245b359bb3117248f7322e