Remove the --autofree option from boot.ipxe
Many distros create and distribute images based on the v1.0.0 code from
iPXE (stable release). This version does not support the --autofree flag
that we use in our boot.ipxe template.

While using --autofree may be a good practice (it frees the image after
chainloading) it's not mandatory and by removing it we will get a much
support for people using the stable version of iPXE and it will continue
to work with the images built from the master.

Closes-Bug: #1486950
Change-Id: Iff936b72aa897c02bd8210cb7687abd255f47e53