Implement take_over for iscsi_ilo driver
When iscsi_ilo driver is using a web server for storing boot images,
an active node is tied down to a specific conductor running the
webserver.  When it is taken over by a different conductor, the boot
image needs to be recreated on the new conductor when it takes over
the node.

Implements: blueprint remove-swift-dependency-for-ilo-drivers

Change-Id: I70cb8e7d624c8a91b4aa2bb550a5f4b6f64cb2b6