Reset tempdir config option after NestedTempfile fixture applied
When running unit tests, tempdir config option is set to
tempfile.gettempdir() by default in ironic/common/utils, which
happens before fixtures.NestedTempfile() is applied, which causes it
to be a system tempdir. Reset its value after fixture is applied so
that nested directory would be used as 'root' directory for temporary
files.

Change-Id: I1e717c6e7b421ff212f3ef3d30577d24362e1c44
Closes-bug: #1489796