Handle missing is_whole_disk_image in pxe._build_pxe_config_options
This commit handles the situation when 'is_whole_disk_image'
is not set in node.driver_internal_info and kernel/ramdisk
is missing in node.instance_info when
pxe._build_pxe_config_options is called.  This can happen when
prepare_ramdisk() is called on a node managed by agent_* driver
that is transitioning from 'manage' to 'available' state for the
first time (after being enrolled).

Change-Id: I6cac1beadc9177dd0cc28bdff7f30a6b3900ec98
Closes-Bug: 1492998