Refactor deploy_utils methods
deploy_utils.agent_get_clean_steps() is generalized so that by default,
it returns clean steps for all interfaces. To specify a particular
interface, an 'interface' argument is added to the method.

Also override_priorities argument is added, so that we can change
agent-provided priorities to ones from configuration.

Change-Id: I529d99e5072074ecbe175f4558d37e4a6023e76a
Co-Authored-By: Ramakrishnan G <rameshg87@gmail.com>
Co-Authored-By: Dmitry Tantsur <dtantsur@redhat.com>
Implements: blueprint inband-raid-configuration