Allow tftpd usage of '--secure' by using symlinks
Create symbolic links in the tftp_root with relative paths. This is to
allow the tftpd server to be run in '--secure' mode.  '--secure' mode
has the daemon doing a chroot upon startup into the specified tftp_root
directory.

Absolute path symbolic links break the use of the '--secure' option.
Using a 'map-file' does not fix the breakage.

For example, previously Ironic might create the file:
     /var/lib/tftpboot/pxelinux.cfg/01-b8-ae-ed-71-18-52 which is a
symbolic link to:
  -> /var/lib/tftpboot/cf15449c-7d09-4b6c-b2db-7e07795ebc98/config

It will now use relative symbolic links so that:
     /var/lib/tftpboot/pxelinux.cfg/01-b8-ae-ed-71-18-52 is a symbolic
link to:
  -> ../cf15449c-7d09-4b6c-b2db-7e07795ebc98/config

This change will affect MAC address linking, for example the symbolic
link:
  /var/lib/tftpboot/pxelinux.cfg/01-b8-ae-ed-71-18-52
also IP address linking, for example the symbolic link:
  /var/lib/tftpboot/10.0.0.1.conf

Change-Id: I0f65030fe044aa19294f5389dbdb829ede61032a
Closes-Bug: #1494991