Fix iBoot test__switch_retries test to not waste time sleeping
The test "test__switch_retries" for the iBoot driver takes more than 2
seconds to run, it's spending time sleeping between one retry and
another. This patch fix the problem by setting the "retry_interval" to 0
seconds. To be safe, the "max_retries" is also set to 0 on the SetUp()
method of the test class, testing willing to verify the retry can
overwrite those within the test function scope.

Closes-Bug: #1496026
Change-Id: Ic735baa532353b4ab0bec0c704b5c11b612433fd