Fix a bug error by passwords only includes numbers
When an IPMI password consists only of numbers, Ironic fails to create a
password file. From the format point of view, these passwords are
correct. These are that it is stored to the variable after being
converted to an int type, cause errors in the subsequent processing to
expect that it is a string type.

Change-Id: I85ee2e904fdb0d538f69c7dbfab4c81bb080b8cf
Closes-Bug: #1489234
(cherry picked from commit 93f99381acd727f539d3a78c0ae82b95860ba649)