Use power manager to reboot in agent deployments
This patch updates the reboot_and_finish_deploy function
for IPA based deployments so that we use the power manager
to send a proper REBOOT command to the instance after
trying to soft power off the client. This is important
for machines that use "external" power controllers which
may not know the power state of the host machine. In
such configurations users may rely on BIOS boot on power on
settings to active their machines. Simply sending a POWER_ON
in such configurations doesn't work because the machine is
already in a POWER_ON state (because it just deployed) so
nothing happens (the machined has power but isn't booted).

Change-Id: If5deddc1ecb069b68f2f1d766eb058691ca280af
Closes-bug: #1503839