Add an iboot reboot_delay setting
This patch adds a new reboot_delay setting to the iboot
power driver. This setting is important when using IPA
which can cause a power manager REBOOT which if there
is no sleep can occur so quickly the machine doesn't properly
power cycle.

The patch also updates the ironic.conf.sample file.

Closes-bug: #1504258

Change-Id: I70ec7151d0154bd6f6445bcc0e90de2dc5b380e5