iPXE: Fix assumption that ${mac} is the MAC of the NIC it's booting
The ${mac} variable not necessarily points to the MAC address of the NIC
we are booting from. This patch updates the boot.ipxe script to make it
more robust by interating over all NICs found in the machine and boot
from the first one capable of booting (the one that contains a
configuration file in the server). If no NICs are capable of booting the
script will fail just like before telling the user that the network boot
failed because we couldn't find any configuration for the present NICs.

In case the inc command is not found, the boot.ipxe will fall back to
the old behavior.

This backport is the combination of 2 commits (squashed):

(cherry picked from commit 1c28eb79a5d1b25fb685e2f48449c51e905766b5)
(cherry picked from commit f968e1a22d3ed928278a04f0da9124678474ff0a)

We need to squash both patches because they must merge at the same time
to avoid breakages.

Closes-Bug: #1507738
Closes-Bug: #1504482
Change-Id: I5b8f0740fa42f611e84b66dc7f6578cffe02760a