Add reboot_delay option to snmp driver
When node is powered off via "soft off" (during deploy with agent for
example) a off/on cycle is required on external power controller.
If pause between off and on is too short node's hardware unable
to detect external power loss.
This patch adds new 'reboot_delay' config option. Some of
SNMP-enabled power controller have own value for this pause
(configurable or hardcoded). So default value of Ironic option
is "0" for keeping of current behavior.

Closes-Bug: #1506860
Change-Id: I91617f1a296ad38ff4c31872e6e3ac20f8619a38