Use server_profile_template_uri at scheduling
Since Server Profile Templates define configuration options that can change
important characteristics of a node, it is important to make this property
available in Nova flavors. That way the scheduler selects a node with accurate
configuration.

This patch moves the server_profile_template_uri from driver_info to
properties/capabilities so the user can use a properly defined flavor to
get a node with an appropriate configuration.

Co-Authored-By: Gabriel Assis Bezerra <gabrielb@lsd.ufcg.edu.br>

Change-Id: I543a07c8978fae4adc29ccd23e7943429aa3186a