Rename utils.py to mgr_utils.py to avoid namespace collision
The file ironic/tests/unit/conductor/utils.py is used during testing but
there is also an ironic/conductor/utils.py and then it is confusing
which test_* file belongs to which file.

Rename utils.py to mgr_utils.py and rename the test file

Intentionally not renaming test_conductor_utils.py to test_utils.py at
this time as we would lose git history for the files. This is best done
in a follow-up patch.

Updated all import statements for unit tests using this file.

Change-Id: I5dfa1ffc01c5f54dc41a7caed83cd0bca3fb0be6