take_over() for iscsi_ilo doesnt recreate boot iso
When using web server for hosting images for ilo driver,
take_over should recreate the boot ISO, if it was created
by Ironic. If the boot iso is not recreated, then the
node doesn't boot up successfully when a node is powered off
and powered on.

The fix is available in master branch via
I5481e705892f66cb3184ac16ff56c123f5c69087 in file boot.py
from lines 117 to 122. The test case for the fix is available via
https://review.openstack.org/#/c/239627 in master branch.
The master branch patch I5481e705892f66cb3184ac16ff56c123f5c69087
implements a feature (~2K lines of change) while the fix for this
issue is just 5-6 lines of change.

Closes-bug: #1510425

Change-Id: I6b7f3901f773a5aba906c50dcbd574412f7b6af3