Sort root device hints when parsing
In order to have a predicable string to test function parsing the root
device hints we need to sort the list. I understand that this is
changing code to help with tests but tests are important and since root
device hints is just a handful of hints it should have no performance
impact in the code.

Change-Id: I7211cab3af9e9974a6ea92fc8933b73ef842829b