Specifying target provision states in fsm
In ironic, a state can be associated with a target state, and these
associations are defined when the fsm is created. This patch adds
the ability to override these default target states when initializing
the fsm as well as when processing an event that transitions the fsm
to a new state (and target state).

This is useful, for example, when adding the manual clean feature,
which causes a node to go from manageable -> clean* -> manageble
instead of the default behaviour (for automated cleaning) of
-> clean* -> available.

Change-Id: I0a7c9be0e44b75f8bd2ebfefd3fb710109612578
Implements: blueprint manual-cleaning