Support arguments for clean step methods
With this change, a clean step method used for manual cleaning can take
keyword arguments.

This adds an optional argsinfo parameter to the clean_step decorator.
Its value is a dictionary of keyword arguments where key is the name
of the argument and value is a dictionary as follows:
- ‘description’: <description>. This should include possible values.
- ‘required’: Boolean. True if this argument is required. If so, it
  must be specified in the clean request; false if it is optional.

Change-Id: I3bcf1a5a91f06b19d917be2bf7acb035801f9193
Implements: blueprint manual-cleaning