Fix: Next cleaning hangs if the previous cleaning was aborted
When cleaning is abort Ironic leaves the node's clean_step field
populated with the step that was running at the moment. Because of that,
once the next cleaning starts again (after the node have been moved to
"manageable" and in the transition to "available") it will hang and
never finish, the reason for that is because once the node starts
heartbeating again Ironic won't know if the cleaning just started or
it's just continuing cleaning (it uses clean_step as a flag for it,
see the bug linked with this patch).

This patch fix this by cleaning up the node's clean_field once the
cleaning process is aborted.

Closes-Bug: #1518374
Change-Id: I423e89f9574a8952875ad9f03ce33298cb5660e7