Add a developer FAQ
This is a first stab a developer FAQ in our documentation. The only
question answered so far is "How do I create a migration script
template?" because every time I have to google that.

This FAQ should be extended with more answers from questions we get from
day to day.

Change-Id: I3a81261892c359e624ed2dbeb7cd42b11b799a6d