Fix iPXE template for whole disk image
The goto :boot_whole_disk section in the iPXE template is bogus, the
"kernel" command is being used to chain the "chain.c32" file, that
command should be "chain" and the "append" command is not a valid iPXE
command.

This patch replaces those commands with the "sanboot" command from iPXE
which is (among other things) used to chain the boot process to the
local disk.

Closes-Bug: #1524403
Change-Id: I7d2310d94e40b1c9c76957382ab19c0076bf42cd
(cherry picked from commit 952d5c5a06f35181768e52dd660afe548bc6fa60)