Add support for dib based agent ramdisk in lib/ironic
This commit adds support for building and configuring
dib based agent ramdisk for Ironic. The environment
variable IRONIC_RAMDISK_TYPE can be used to
mention the preferred distro to build the ramdisk.
The supported types are: coreos, tinyipa and dib.
By default, lib/ironic will continue building coreos
based IPA ramdisk.

The environment variable IRONIC_DIB_RAMDISK_OPTIONS
specifies the options for disk-image-create command
and the platform on which to build the dib based
ironic-python-agent ramdisk.
Default value for IRONIC_DIB_RAMDISK_OPTIONS
is 'ubuntu' when 'dib' is the selected distro.

This patch is required for iLO ThirdParty CI.

Closes-Bug: 1510797

Change-Id: I0b7f45b87f92b739c0926f2de6c134f9d4e426bb