Fix issues with uefi-ipxe booting
When booting from ipxe in uefi boot mode, once the node boots with deploy
image, ironic is not changing the default menu entry to boot_partition
to boot the instance image during next boot. The current changes fix
the issue. The kernel parameter "initrd=ramdisk" is also added to the ipxe
template which is required for ipxe booting.

Closes-Bug: #1532679
Change-Id: Idf78195d228b4861f85fbcae9a269edd95523ad7