Active Node Creation via adopt state
At present the ironic API explicitly sets the new state for
nodes to the beginning step in the ironic workflow.

As part of hardware fleet lifecycle management, an operator
expects to be able to migrate inventory and control systems
for their hardware fleet utilizing their existing inventory
data and allocation records. Ultimately this means that
an imported host MAY already be allocated and unavailable
for immediate allocation.

As such, a mechanism is required to permit users to put nodes
into an ACTIVE state without performing a deployment operation.

This adds a new API provision_state verb to allow users to move
nodes from MANAGEABLE state to ACTIVE state.

Partial-Bug: #1526315
Change-Id: Ib3eadf4172e93add9a9855582f56cbb3707f3d39
Depends-On: Ie114bfaab249d73ea3ca7c0edc314ca1ed0448eb