Remove conditional checking the auth_strategy values
In the dhcp/neutron.py module there was a conditional checking if
the values of the CONF.neutron.auth_strategy was set to "noauth" or
"keystone". That check is not needed anymore since that configuration
option is now using the "choices" parameter from oslo.config to limit
and check what are the valid values for that configuration option.

Change-Id: I61c5e83b06a370bbc4cca48c8812f9b4b934aef8