[DevStack] fix restart of nova compute
We introduced nova compute restart to handle problems of it failure
to start when keystone is not available. However, now we get broken
again because stop_service fails when nova compute is not running.
This change makes it ignore failure to stop nova compute.

Also, if n-cpu failed to start the first time, the n-cpu.failure file
that
was created will fail the service_check function in the end
of the deployment, although n-cpu might be in fact up, running and
healthy.

Rename this file to `n-cpu.failure.before-restart-by-ironic` right after
stopping n-cpu so that it does not fail service_check later, but is
still
available for reference if needed.

Change-Id: Idfe7d68d6ebeb786fe9e6da6a001a0ee2e10a76b
Closes-Bug: #1546058
Related-Bug: #1537076