Add links to the standalone configdrive documentation
Prior to this patch the standalone section of the configuration drive
documentation was talking about a specific format required to the
configuration data but there was no link to the said format, that was
confusing. This patch is adding a link a documentation with an example
of how the data should be organized.

Change-Id: I2fb006c79a20a4795465ce1d992ce8103e6a909b