add wipefs to ironic-lib.filters
I39637e22c344703ad48fc271f6f866aa018bbdd1 introduces the use of wipefs
to clean a disk, this through the ironic_lib.disk_utils module. wipefs
has been added to ironic_lib's rootwrap filter filter as expected.

We need to keep this one in sync because:
- of the gate: devstack uses it instead of the one from ironic_lib
- to preserve exsisting deployed installation where ironic_lib's filter
  is not properly deployed yet.

This commit also comes with a NOTE message to explain the situation.

Change-Id: Id69706d7a78ff37a360cc23d6a076bc13bbdf14c