Add documentation for the IPMITool driver
This patch is adding the documentation for the IPMITool driver. This
covers some general overview, basic setup (prerequisites, how to enable)
as well as some advanced topics such as single/double bridge and protocol
version.

The IPMITool driver is rich and allows a lot of configuration, therefore
not all topics are addressed by this commit. TODOs were left in the
documentation (as comments) so we don't forget to address them later.

Partial-Bug: #1468710
Change-Id: Ife99de75a42d7048fd1ad31e220b5606a8fe81f0