DevStack: Allow configuring the authentication strategy
This patch is making the auth_strategy configurable in the DevStack
module. In many occasions I want to use the same user to list the Ironic
API and nova API when debugging the deployment with DevStack, in order
to avoid having to use the admin user to issue Ironic commands we have
to edit the ironic configuration file, set the auth_strategy to "noauth"
and restart the ironic-api service. By making the auth_strategy
configurable we can set it to "noauth" before we start the [re]creation
of our stack and avoid the steps above.

Change-Id: I99c77cb6d86af7938a068f47d670d566fe083ed3