Agent: Out-of-band power off on deploy
The patch is adding  a new property called "deploy_forces_oob_reboot"
for the Agent deployment methodology. If the property is specified in
the node's driver_info with the value of True, Ironic will call IPA to
flush the file system and then issue a out-of-band reboot at the end of
the deployment to the node to boot into the user's image.

Some hardware/firmware may have problems doing a power off in-band
(usually faulty hardware/bad firmware, see the bug this commit is fixing)
therefore we need Ironic to do it directly in the BMC.

This patch also closes one more consistency gap between the IPA deploy
model and the now deprecated bash ramdisk one which used the hard reboot
to finish the deployment by default.

Depends-On: I5cd1d1b821426e995dc584452494b93ab23917e0
Closes-Bug: #1512492
Change-Id: I5e389c5245826f86466aedfcb9730117e24b59db