[devstack] Do not die if neutron is disabled
One of the recent changes added a die_if_not_set statement in
init_ironic if the script is not able to get cleaning_network_uuid
in neutron. This makes ironicclient-functional job to fail, as it
defines OVERRIDE_ENABLED_SERVICES=key,mysql,rabbit,ir-api,ir-cond.
We should die_if_not_set only when neutron is enabled.

Closes-bug: #1557709

Change-Id: I9f8f38d0fdf058417af31f44e86fff099c251082