Allow user to specify cleaning network
Currently devstack assumes that the cleaning network will be
the private network. This prevents the use of provider and
public networks to be used for the cleaning network, which is
useful when deploying to actual hardware. This patch adds
the IRONIC_CLEAN_NET_NAME var which allows other networks to
be specified.

Change-Id: I5006dbe9f1cf674e5df4baad6646d827cf362c55