Add user and project domains to ironic context
This change also removes most of the logic from ironic's
RequestContext to reuse the oslo_context as much as possible.
Usage of domain_id and domain_name in policy files is deprecated
and their support will be removed in the Pike release. domain_id
field was removed from the context class completely, domain_name
value now mathces the oslo_context expectations.

ContextHook is changed too so as not to duplicate from_environ
functional from oslo_context.

to_dict method left as is, so that we don't break an older service
receiving the context over RPC. It will be changed in Pike release
to reuse the base oslo_context class' to_dict.

Closes-Bug: #1602081
Closes-Bug: #1627173
Closes-Bug: #1641972
Co-Authored-By: Jamie Lennox <jamielennox@gmail.com>
Co-Authored-By: Devananda van der Veen <devananda.vdv@gmail.com>
Change-Id: I9afe89bc6aee282ee4b7579d661e3fa83cc0ce84