Don't power off non-deploying iLO nodes in takeover
The iLO driver, when the conductor takeover logic was triggered
would power off the node if it was not in the ACTIVE state,
which could potentially be harmful to the hardware.

Presently, takeover is only called for active nodes, however in
the future, features like ADOPTION will leverage take_over and
evolution of redundancy support may drive additional states that
the node is in when takeover is called.

This changes the behavior and unit tests accordingly such
that only nodes in deploying state perform deployment related
method calls and thus prevent accdiental actions occuring.

Change-Id: I244c3f31d0ad26194887cfb9b79f96b5111296c6
Closes-Bug: #1563949