Extend IRONIC_RAMDISK_TYPE to support 'dib'
This fix avoids the devstack to stop proceeding if
the setting IRONIC_RAMDISK_TYPE is not coreos or tinyipa.
The change ID I641a5afe8318ec0848395b2b70e056c448f3dd00
forces the stack to stop if IRONIC_RAMDISK_TYPE is other
than coreos and tinyipa. There is no tar ball which can be
fetched if the IRONIC_RAMDISK_TYPE is dib.

Closes-Bug: 1567013
Change-Id: I74906e83f7e780d119302b47cf542f22a535c966