Add note on prerequisite of 'rpm' file extraction
As part of the manual cleaning step of firmware update the extraction
of firmware file with .rpm extension fails if commands `rpm2cpio` and
`cpio` are not present on the conductor. This doc note adds the
prerequisite information about `rpm2cpio` and `cpio` tools while using
a .rpm extn firmware file.

Partial-Bug: #1569900
Change-Id: I806980e62a9ad699b5a278a9dbc07682115d2a8f