File download fails with swift pseudo folder
As part of manual cleaning step of firmware update the download of
firmware file fails if the firmware file is associated with swift
pseudo folder. Currently only basename of the parsed swift url path
is taken into consideration while retrieving the object name from
swift. This is to fix that by now taking the complete path of the
swift object to include the pseudo folder as well, if included.

Closes-Bug: #1571496
Change-Id: Ib71692e191501701a92cd7acacb172d0d7fce922