FIX: IPMI bmc_reset() always executed as "warm"
When the invoked via the CLI/API the bmc_reset() "warm" parameter will
receive a string, prior to this patch the string would be seem as True
(since the parameter was expecting a boolen) causing the driver to
always reset the BMC in "warm" mode.

This patch fixes the problem by converting the string passed to boolean
inside the method.

Change-Id: I052ef1ee911703c9b7316362aced17b6303b991d
Closes-Bug: #1573614