Make the ssh driver work on headless VirtualBox machines
On a workstation where an X display might not currently be running
or a workstation where someone else is logged on the graphical
desktop this small change helps get things working for using
VirtualBox as the hypervisor. Use VirtualBox 3.2 or above
for headless mode support.

This patch also ensures that set_boot_device() always works by powering
off the VM before setting the device. VirtualBox does not allow boot
device to be set while the machine is on.

Closes-Bug: 1561624

Change-Id: If8d989cc3964e849544b139959b329a3fd05d451