Remove backward compat for Liberty cleaning
In Liberty, we fetched the last clean step by comparing the current
clean step to the full list stored in the node. In Mitaka, we cleaned
this up with node.driver_internal_info['clean_step_index']. However, we
left a shim in for nodes that were cleaning during the Liberty to Mitaka
upgrade, and did not have the clean step index stored.

Now that we're in Newton, we can safely remove that shim.

Change-Id: I444cd5188a1dfb56fbad13f025179dfdc4cb5c85