Correct RAID documentation JSON
The original example dictionaries added to the RAID documentation
used single quotes instead of double quotes. Since double quotes
are used in JSON, as opposed to dictionaries, to represent a string,
they were failing JSON validation and could not be directly used.

Changed the single quotes to double quotes where appropriate
in the RAID documentation and passed each example through a JSON
validator which now returns that the example is valid JSON.

This patch also updates quotes in cleaning.rst and ilo.rst.

Change-Id: Icb49dfd6722771c147ff5055a6562b6a6ebecb6a
Closes-Bug: #1586157