Remove support for 'hexraw' iPXE type
This removes support for "hexraw" type in the iPXE script
(boot.ipxe) since "hexraw" is not supported in older
versions of iPXE. "hexhyp" replaced "hexraw" and has been
used since kilo.

Also addresses a nit in the original patch
(https://review.openstack.org/#/c/171759/). For
_get_pxe_mac_path(),  changed the assigned value of
'delimiter' argument from None to '-' (because it
was coded to be '-' as the default).

Change-Id: Ide05ff7c4689cd696e2fe76be98ff99c0be24236