Mask password on agent lookup according to policy
We currently mask passwords in driver_info for all API responses,
except for agent lookups. This is because driver_vendor_passthru
just expects a dict to return as JSON in the response, and doesn't
modify it at all.

Change lookup to follow the defined policy when returning the node
object in the response, the same way we do for other API responses.

Co-authored-by: Jim Rollenhagen <jim@jimrollenhagen.com>

Change-Id: Ib19d2f86d3527333e905fdbf1f09fc3dc8b8c5a7
Closes-Bug: #1572796