Change the logic of selecting image for tests
Right now devstack sets DEFAULT_IMAGE_NAME depending on the driver in
stackrc. UUID of the image with this name is then written to tempest
config during tempest configuration stage and is used to determine
which image will be used during nova boot.

As ironic python agent now supports both whole disk and partition
images, we should change this logic, and the user should be able to
request which image should be used in tests.

This commit adds IRONIC_TEMPEST_WHOLE_DISK_IMAGE and
IRONIC_IMAGE_NAME variables to provide this possibility.

If IRONIC_TEMPEST_WHOLE_DISK_IMAGE is not set, we try to infer the
image type by the size of the ephemeral partition requested, if it
is 0, we consider the image to be whole disk.

Change-Id: Ic718a6b37c09ae564ecc3913cf31234316f1da63
(cherry picked from commit e434225f9bec6f078f5514c82c9406131122a49e)