Change port used for Ironic static http to 3928
Due to recent infrastructure changes, infra began using port 8088. This
caused a conflict with our apache config, causing services to not start.
Infra is changing this port (Ie67bbba02dbf61a481f66001de3e0dede9448316)
but it may take up to a week to take effect.

Instead of having our gate broken for a week, I'm changing the default
port. 3928 was not chosen at random; it's listed as the netboot-pxe
service port in /etc/services.

Change-Id: Ifdcf3c82fd9ee64c1548e47dba4c78c2347959e0
Closes-bug: 1590139