Bug fixes and doc updates for adoption
During the creation of a tempest test for the adoption feature,
some minor issues were identified with the adoption functionality.

Namely, the default logic was to create ramdisks, however that logic
path is more intended for deployment ramdisks. Logic was switched
to the instance preparation logic which is the default for nodes in
active state, which is realistically exactly what is desired.

Validation behavior ultimately remains unchanged and tests were
added to validate that the expected methods are called.

Additionally, it was identified that it would be ideal to
encourage the user to set the node to local boot, and as
such the documentation was updated as part of this change,
coupled with a note mentioning changes in API version 1.20
that a user may wish to leverage.

Change-Id: Id6053e0fa68deb431f4543005421982c795401f2
Closes-Bug: #1605239