Refactor common checks when instantiating the ipmitool classes
The constructor for the ipmitool classes are the same but they are
duplicated all over this patch is refactoring that code into a common
function that can be called from the classes constructors.

The VendorPassthru class wasn't checking for the 'timing' option support
of the ipmitool command, it should, this patch is fixing that too.

Change-Id: I456dbb9ef11230d722ff1b1fe5aa142237e0d187