Add storage_interface to node DB table
Add storage_interface DB field, database migration,
and object entry for storage_interface for boot from
volume functionality.

Initial versions of this revision added the field to
the node object, but since then Ironic's code has
drifted such that adding an object for an interface
has to take place along with the entire driver interface,
thus the object change is in the next revision.

Co-Authored-By: Stephane Miller <stephane@alum.mit.edu>
Change-Id: Ib5d5de87163570949c12d9e95c73e17c1c12eec1
Partial-Bug: #1559691