Add a data migration to fill node.network_interface
This fills node.network_interface based on the default logic that the
node update/create calls follow, so that we know network_interface is
populated when this code is deployed.

Also adds to existing release note about how network_interface is set, to
remind people to check the logic and configs to make sure they get the
expected result.

Change-Id: I09a42c8e54d7782c591415e53fccade972ae8bdb
Closes-Bug: #1608511