Configure clean network to provision network
Currently when setting up Ironic with multi-tenancy support,
the DevStack plugin by default still sets the cleaning network
to the default private network of the demo project,
which is not exactly correct.

Change the DevStack plugin so that the cleaning network
is the same as provisioning network if IRONIC_CLEAN_NET_NAME is not set:

- In case of multi-tenancy support enabled, use the created separate
provisioning network
- w/o multi-tenancy support use the default private network.

If the IRONIC_CLEAN_NET_NAME is provided, use that.

Change-Id: I9db8f5e3000f05d27a3fa8fc94cfdafe34a6f48a