Fix tempest realted exceptions during docs build
We don't have tempest in our {test-}requirements.txt, which causes
docs build to complain about missing tempest.* modules, as we have
autodoc enabled, which imports all the modules it can find to get
their docstrings.

This change adds ironic_tempest_plugin modules to
autodoc_exclude_modules, and also adds it to exclude_patterns, so
that sphinx ignores ironic_tempest_plugin/* files too.

Also fixes a couple of warnings because of the incorrect formatting
of the LocalLinkConnectionType.validate docstring.

Closes-bug: #1609850
Change-Id: Id24e587b690b423e65dad55e70224426873e8d5d