Makes ironic build reproducible
This changes the samples for port, volume_connector, and volume_target
so that the updated_at and created_at values are constant times, similar
to the node and chassis.

This patch makes the build of Ironic reproducible. If you don't know
about reproducible build, please read on:
https://wiki.debian.org/ReproducibleBuilds
https://wiki.debian.org/ReproducibleBuilds/About

Debian has SOURCE_DATE_EPOCH set according to the last date of
debian/changelog, which makes it so that the time is constant,
and the build becomes reproducible.

Change-Id: I302609bf53659e33e89cae223336a453052a5745
Co-Authored-By: Ruby Loo <ruby.loo@intel.com>