IPMITool: Check the boot mode when setting the boot device
This patch is fixing a problem where the ipmitool option "efiboot"
wasn't passed as part of the command to set the boot device for UEFI
nodes causing them to switch back to legacy BIOS.

A FIXME note was left in the code. It seems that ipmitool has a problem
setting the efiboot + persistent options at the same time
(see bug #1611306). It seems multiple options ought to be specified
together, and that works, except for efiboot. Therefore this patch is
using raw bytes to change the boot device when persistent and uefi are
specified, that way it will work for newer and older versions of the
ipmitool utility.

Closes-Bug: #1611306
Change-Id: I11456fb035fe41fe46f1ff04e4d56c9f85b9e19d
(cherry picked from commit 6419cc2118e9d8c613156f6d4aeae5db46aa385d)