Copy iPXE script over only when needed
currently the iPXE boot script is overwritten on every `prepare_ramdisk`
call, which is inefficient.

Instead, copy it over only when it is missing from HTTPROOT or is
different from the one specified in the config.
The comparison is performed using `filecmp` from Python standard lib,
which in turn uses `os.stat()`, so although being not very strict,
comparison does not call to any extra utilities and is fast.

Change-Id: I1213fc9a9945e11f963d10f0581a7d69ef13ee53