Log node uuid rather than id when acquiring node lock
When attempting to get lock of a node, we get such log:
    "Attempting to get exclusive lock on node 2 (for
     port update)"
However, when reserving and releasing the same node lock,
we get such log:
    "Node 7c3c747d-6315-4db8-8fa5-8e55cf2bcc34 successfully
     reserved for port update (took 0.02 seconds)"
    "Successfully released exclusive lock for port update
     on node 7c3c747d-6315-4db8-8fa5-8e55cf2bcc34 (lock
     was held 0.04 sec)"

Sometimes, we log node_id but in other places, we log
node_uuid. It will become a challenge for trouble shooting
with such inconsistent log.

With this change we always log node_uuid when acquiring node
lock.

Change-Id: Icac16d490d654c73cc99e5d2a9a7c682adbc2802
Closes-bug: #1602490