Pass swiftclient header values as strings
Swiftclient passes header values straight through to requests, which
recently started enforcing a long-documented requirement that headers
be byte- or unicode-strings. While the Swift community decides whether
and when to str() header values vs. raising an error [1], let's
explicitly cast what we know to be ints to strs as that should *always*
work.

[1] This is a bit more complicated than it seems at first glance;
    see the discussion around the requests change for details at
    https://github.com/kennethreitz/requests/pull/3366

Change-Id: Ic52a68365dc08db055f776ec8d48f7bf44d6b80d
Related-Bug: #1614932