Allow suppressing ramdisk logs collection
A third-party deployment driver that is not IPA-based may still use
'set_failed_state' method from 'deploy_utils', but currently this will
result in an ERROR-level log entry, unnecessary worrying the operator.

Allow calls to this method to suppress IPA-based ramdisk logs collection
even when it is enabled in config.

Change-Id: I4d9fe0382c012d6a00826812752ffe65b5a98ea3