OneView test nodes to use dynamic allocation
Now that the dynamic allocation feature has merged, the testing of *_oneview
drivers shall use the new allocation model on the CI infra. To allow this, we're
adding a new parameter to the hwinfo file to indicate when the
dynamic_allocation flag should be added to the node and managing the logic when
it is present.

Change-Id: I40ed9abe96dd15701b31e0ad0b7d71ff065488ef