[install-guide] Import "Bare Metal service overview"
Import service overview in the new install-guide and provide a link in
the old guide that points to it.

Additionally, Swift was added to the list of interactions and Cinder was
removed since it is not yet available.

Change-Id: Ica6a9aae2c35f293228ffb75b2cf82478eb568a5
Partial-bug: #1612278
(cherry picked from commit 7131037cf7665680adee143a0f886e2c7ddd0327)