Increase verbosity of devstack/lib/ironic
Increase the verbosity of devstack/lib/ironic by enabling xtrace for
when sourcing devstack/lib/ironic. This information is useful to be able
to see what variables are set to along with other information.  This
will increase the devstack log size by about 1%.

Also print which stage the ironic devstack plugin is being called for.

Change-Id: Ibd2530720f21c0b0c03af5c7ed2a6d7aa68eda65