Fix broken unit tests for get_ilo_object
First, the tested function signature was wrong. We didn't catch it in gate,
as we mock proliantutils, but it does break e.g. Debian package build.

Second, the arguments override was not actually working. We didn't catch
it in gate, because the new values were the same as the defaults.

Closes-Bug: #1626089
Change-Id: I2e4899e368b0b882dcd59bf33fdca98f47e5b405