Fix DRAC passthru 'list_unfinished_jobs' desc
This fixes the description for DRAC's "list_unfinished_jobs()"
passthru method. It incorrectly indicates that there are
required arguments, but no arguments are required from the caller
of that REST API. The description indicates what is returned
by that method.

This is a followup to I963504dbbec36e44312ccbf4455c45d6ec60908b.

Change-Id: I4070f30b9898ef9f17c585ddba68cdac4d6335c7
Related-Bug: #1621515