[install-guide] describe service clients auth
Currently install guide incorrectly advises operators to configure
access for service users used by ironic-conductor in
[keystone_authtoken] section.

Such way of configuration was deprecated in Newton release, and using it
will result in a warning log message.

This patch changes the advised way of configuring service users to an
appropriate one using config sections corresponding to each service client
to be used, as well as proper options for keystoneauth plugins and
sessions instead of legacy options.

Change-Id: I96967682f18988d9e32f2ecfa7792aa0e653db3f
Closes-Bug: #1628842